pyproject.toml
[project]
name = "daily-twilio-sip-dial-in"
version = "0.1.0"
description = "Daily SIP + Twilio Dial-in example"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pipecat-ai[daily,webrtc,deepgram,cartesia,openai,silero,runner,local-smart-turn-v3,remote-smart-turn]==0.0.85",
    "pipecatcloud>=0.2.4",
    "twilio==9.8.0",
    "sentence-transformers==5.1.0",
    "qdrant-client==1.15.1",
    "python-dotenv>=1.0.1,<2.0.0",
]

[dependency-groups]
dev = [
    "pre-commit~=4.2.0",
    "ruff~=0.12.1",
]

[tool.ruff]
exclude = [".git", "*_pb2.py"]
line-length = 100

[tool.ruff.lint]
select = ["I"]
ignore = []